+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/config/group/EndpointGroupingRule.java
@@ -53,4 +53,8 @@ public class EndpointGroupingRule {
             return new StringFormatGroup.FormatResult(false, endpointName, endpointName);
         }
     }
+
+    void setRules(String serviceName, StringFormatGroup formatGroup) {
+        rules.put(serviceName, formatGroup);
+    }
 }